+++ config/environments/development.rb
@@ -1,6 +1,6 @@
 require 'active_support/core_ext/integer/time'
 
-Rails.application.routes.default_url_options = { host: 'era.ualberta.localhost' }
+Rails.application.routes.default_url_options = { host: 'era.lvh.me', port: 3000 }
 
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
@@ -33,7 +33,7 @@ Rails.application.configure do
   end
 
   # Store uploaded files on the local file system (see config/storage.yml for options).
-  config.active_storage.service = (ENV['ACTIVE_STORAGE_SERVICE'] || :local).to_sym
+  config.active_storage.service = (ENV.fetch('ACTIVE_STORAGE_SERVICE', nil) || :local).to_sym
 
   # Don't care if the mailer can't send.
   config.action_mailer.raise_delivery_errors = false