+++ test/jobs/digitization/batch_metadata_ingestion_job_test.rb
@@ -15,7 +15,7 @@ class Digitization::BatchMetadataIngestionJobTest < ActiveJob::TestCase
       end
     end
 
-    assert(batch_ingest.failed?)
+    assert_predicate(batch_ingest, :failed?)
     assert_equal('Testing! Error has happened!', batch_ingest.error_message)
   end
 
@@ -29,7 +29,7 @@ class Digitization::BatchMetadataIngestionJobTest < ActiveJob::TestCase
     end
 
     batch_ingest.reload
-    assert(batch_ingest.completed?)
+    assert_predicate(batch_ingest, :completed?)
     assert_equal(2, batch_ingest.books.count)
   end